Correction of path for buildspec generated file
Prior to version 0.4, the generated buildspec.yaml file was placed in the directory 'terraform/build'. Due to moving the module files to the root directory, the buildspec file was now generated at the root of the module path, which was unexpected. This has been corrected so that the file is generated at the same path it was before.